requirements.txt
discord.py==2.3.2
yt-dlp==2023.11.16
python-dotenv==1.0.0